refactor: reorganize and rename PostgresLiquibaseTestcontainersExtension
* Group tests into @Nested classes for better structure.
* Rename test class to reflect broader unit testing scope.
* Add new test to verify afterAll behavior.
